# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Repaire Phone DZ - E-commerce API for phone repair equipment
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
        - name: categoryId
          in: query
          schema:
            type: ["number", "null"]
        - name: brandId
          in: query
          schema:
            type: ["number", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: inStock
          in: query
          schema:
            type: ["boolean", "null"]
        - name: isNew
          in: query
          schema:
            type: ["boolean", "null"]
        - name: hasDiscount
          in: query
          schema:
            type: ["boolean", "null"]
        - name: isFeatured
          in: query
          schema:
            type: ["boolean", "null"]
        - name: sortBy
          in: query
          schema:
            type: ["string", "null"]
            enum: ["price_asc", "price_desc", "newest", "popular", "rating", null]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/search-suggestions:
    get:
      operationId: getSearchSuggestions
      tags: [products]
      summary: Get instant search suggestions
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchSuggestion"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # ─── REVIEWS ────────────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listMyOrders
      tags: [orders]
      summary: Get current user's orders
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── COUPONS ─────────────────────────────────────────────────────────────────
...
  /admin/orders:
    get:
      operationId: listAllOrders
      tags: [admin]
      summary: List all orders (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /admin/orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH ──────────────────────────────────────────────────────────────────
...
    Product:
      type: object
      required: [id, name, slug, price, stock, isNew, isFeatured, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        slug:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        discountPercent:
          type: ["number", "null"]
        stock:
          type: number
        sku:
          type: ["string", "null"]
        isNew:
          type: boolean
        isFeatured:
          type: boolean
        hasDiscount:
          type: boolean
        averageRating:
          type: number
        reviewCount:
          type: number
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["number", "null"]
        categoryName:
          type: ["string", "null"]
        brandId:
          type: ["number", "null"]
        brandName:
          type: ["string", "null"]
        createdAt:
          type: string
    ProductDetail:
      type: object
      required: [id, name, slug, price, stock, isNew, isFeatured, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        discountPercent:
          type: ["number", "null"]
        stock:
          type: number
        sku:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        isNew:
          type: boolean
        isFeatured:
          type: boolean
        hasDiscount:
          type: boolean
        specifications:
          type: object
          additionalProperties: true
        shippingInfo:
          type: ["string", "null"]
        warrantyInfo:
          type: ["string", "null"]
        averageRating:
          type: number
        reviewCount:
          type: number
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["number", "null"]
        categoryName:
          type: ["string", "null"]
        brandId:
          type: ["number", "null"]
        brandName:
          type: ["string", "null"]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, price, stock]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: number
        stock:
          type: number
        sku:
          type: string
        barcode:
          type: string
        isNew:
          type: boolean
        isFeatured:
          type: boolean
        hasDiscount:
          type: boolean
        specifications:
          type: object
          additionalProperties: true
        shippingInfo:
          type: string
        warrantyInfo:
          type: string
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["number", "null"]
        brandId:
          type: ["number", "null"]
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: number
        stock:
          type: number
        sku:
          type: string
        isNew:
          type: boolean
        isFeatured:
          type: boolean
        hasDiscount:
          type: boolean
        specifications:
          type: object
          additionalProperties: true
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["number", "null"]
        brandId:
          type: ["number", "null"]
    ProductListResponse:
      type: object
      required: [products, total, page, totalPages]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
    SearchSuggestion:
      type: object
      required: [id, name, type]
      properties:
        id:
          type: number
        name:
          type: string
        type:
          type: string
          enum: [product, category, brand]
        imageUrl:
          type: ["string", "null"]
        price:
          type: ["number", "null"]

    # ── REVIEW ────────────────────────────────────────────────────────────────
...
    OrderItem:
      type: object
      required: [productId, name, price, quantity]
      properties:
        productId:
          type: number
        name:
          type: string
        price:
          type: number
        quantity:
          type: number
        images:
          type: array
          items:
            type: string
    Order:
      type: object
      required: [id, status, subtotal, total, createdAt]
      properties:
        id:
          type: number
        userId:
          type: number
        userName:
          type: ["string", "null"]
        userEmail:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        discount:
          type: number
        couponCode:
          type: ["string", "null"]
        shipping:
          type: number
        total:
          type: number
        shippingAddress:
          $ref: "#/components/schemas/Address"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    OrderInput:
      type: object
      required: [shippingAddress]
      properties:
        shippingAddress:
          $ref: "#/components/schemas/Address"
        notes:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
    OrderListResponse:
      type: object
      required: [orders, total, page, totalPages]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
    Address:
      type: object
      required: [fullName, phone, wilaya, address]
      properties:
        fullName:
          type: string
        phone:
          type: string
        wilaya:
          type: string
        commune:
          type: string
        address:
          type: string

    # ── COUPON ────────────────────────────────────────────────────────────────
...