# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Repaire Phone DZ - E-commerce API for phone repair equipment
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputSpecifications } from './productInputSpecifications';

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  comparePrice?: number;
  stock: number;
  sku?: string;
  barcode?: string;
  isNew?: boolean;
  isFeatured?: boolean;
  hasDiscount?: boolean;
  specifications?: ProductInputSpecifications;
  shippingInfo?: string;
  warrantyInfo?: string;
  images?: string[];
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  brandId?: number | null;
}